extract Services interface
