+++ src/App.tsx
@@ -5,8 +5,9 @@ import SessionContext from 'auth/SessionContext'
 import ErrorBoundary from 'common/ErrorBoundary'
 
 import { useAuthentication } from 'auth/Auth'
-import Router, { Services } from 'router/router'
+import Router from 'router/router'
 import { DictionaryContext } from 'dictionary/ui/dictionary-context'
+import Services from 'router/Services'
 
 if (process.env.REACT_APP_GA_TRACKING_ID) {
   ReactGA.initialize(process.env.REACT_APP_GA_TRACKING_ID)
