+++ src/router/sitemap.tsx
@@ -3,7 +3,8 @@ import DynamicSitemap from 'react-dynamic-sitemap'
 import { renderToString } from 'react-dom/server'
 import $ from 'jquery'
 import { Route } from 'react-router-dom'
-import { Services, WebsiteRoutes } from 'router/router'
+import { WebsiteRoutes } from 'router/router'
+import Services from 'router/Services'
 import withData from 'http/withData'
 import Bluebird from 'bluebird'
 import convert from 'xml-js'
